MCP Project 2/requirements.txt
sqlite3
openai
openai-agents
polygon-api-client
